Rework to use a booted daemon rather than directly
