+++ .gitignore
@@ -1,2 +1,4 @@
 gogo.jar
 gogo.jar.pack.gz
+classes
+gogo-daemon.jar
