--- classes/.gitignore
@@ -1,1 +1,0 @@
-gogoHID
