+++ resources/darwin/libhidapi.dylib
