+++ resources/libhidapipi.so
