+++ resources/linux-x86-64/libhidapi.so
