+++ resources/linux-x86/libhidapi.so
