+++ resources/win32-x86/hidapi.dll
