+++ resources/wind32-x86-64/hidapi.dll
